LAB 10 Week Array
This week is upgrade week 07 (not 06 or 05 in the fucking labsheet) the change file is Customer 1st is the comment one , 2nd is the acct ArrayList. The NEWEST CLASS! Bank to ez (i almost forget the most important part that i often forget. you must import java.uitl everytime you use ArrayList. also <...> <-- this generic one to change class)